ENGR00163128 [MX6q_ARM2]SATA Boot: reboot failed if boot from SATA
The root cause is that the sata power is not enabled after reboot
the system when boot from sata.
Built in the sata driver would fix this issue.
In this way, the sata power would be enabled defautly
when there is a sata device is plugged into the slot, otherwise,
the power would be disabled.
NOTE:
Pls don't plug the sata disk after the system is booted up.
System wouldn't recognized it, since there are no any power and
clocks on the sata slot.

Signed-off-by: Richard Zhu <r65037@freescale.com>